requirements.txt
accelerate==1.2.1
albumentations
attrs
cffi
click==8.1.8
comet_ml==3.49.0
diffusers==0.33.0
dotenv
entrypoints
fastjsonschema==2.21.1
future==1.0.0
huggingface-hub[cli]==0.29.1
ImageHash==4.3.1
jeepney==0.9.0
joblib==1.4.2
jsonschema
loguru==0.7.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
numpy==1.26.3
opencv-python==4.11.0.86
pandas==2.2.3
pandas-stubs==2.2.3.241126
pillow==10.2.0
protobuf==6.30.2
pytest==8.3.4
PytorchWildlife==1.2.0
regex==2024.11.6
requests==2.32.3
seaborn==0.13.2
shellingham==1.5.4
SecretStorage==3.3.3
sentencepiece==0.2.0
scikit-learn==1.5.0
tifffile==2024.5.22
tiktoken==0.8.0
timm==1.0.15
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
verovio==4.5.1
-e .
